expandable file structure setup
more for Redux in future
